Add 2 processes for M_InOut -> M_ShipperTransportation
https://github.com/metasfresh/metasfresh/issues/5885
